refactor: extract compression types and permission decision to turn-core
- Extract TokenBudget, CompressionResult, PipelineOutcome, CompressionLayer trait
  to new compression_types.rs in astra-turn-core
- Keep CompressionPipeline struct in runtime (has runtime-specific factory methods)
- Extract PermissionDecision enum + PermissionCallback type alias to turn-core
- Remove old definitions from runtime, add re-exports
